locally compute line range start
stop computing outside of helpers, compute within the helpers, locally. this frees up the line comment calculation to work differently than the block comment, so that we can start addressing its bad calculation. this also moves work out of the main loop which should not be concerned with the bookkeeping work. this is also one step towards declaring comment context in the outer comment loop instead of the inner line loop in the comment rule function.