revise trailing space handling in block overflow
previously we did not bother to check whether the comment line had a bunch of extra trailing whitespace. it was possible for it to overflow simply because of this
extra whitespace. this is not ordinary because i assume most authors use no-trailing-spaces. however, we cannot rely on that rule being set up. so, we have to trim the trailing whitespace and double check the length. if, after trimming, the remaining content does not reach the threshold, then this should not be considered overflow.

* progress on #23
* progress on #17