begin refactoring block underflow
working on updating block underflow to work similar to block overflow. i have to stop here though, because it turns out that there is a lot of overlap in the preprocessing. there is a lot of copy/paste going in. the functions basically perform some of the same calculations at the start. so i think it makes sense to not continue revising the block underflow in a vacuum. instead, we should revise block overflow to work with shared state, and then revise block underflow to both work differently and work with that shared state.