fix block overcheck for for last line of block
when evaluating the length of the last line of a block comment, we need to consider the suffix, which is the whitespace trailing the content and the final star-slash. for the last line only, that trailing text counts towards overflow.

to facilitate this i added a new suffix property to line data