drop tracking of comment index
the comment index was previously used by the line underflow check to determine if two comments are sequential. that check was changed and is now done using line number. as a result there is no need to track the comment index because nothing uses it. as a result the comment iteration loop is now a simpler for..of.