remove type folders
plan to possibly merge some functionality together and not organize it separately by type. having these type folders around is just making it annoying to start on those changes because type is no longer significant. so one step away from the communicative-coherency and towards the functional-coherency is doing away with the type folders