remove explicit mention of shebang in index.ts
so long as we check type is Block and then check type is Line, and not just check type is Block and assume the only alternative is Line, then we handle shebangs. i do not feel like the commentary adds much help so just remove all mention of it. the only confusing and unstated thing that remains is why we check comment.type is Line if we know it is not type Block.