ensure final line of block comment checked for split
fix an error in the code that iterates over block lines. it was only checking the final line if the number of lines was odd. whoops. the new code unconditionally calls split on the final line of the block. the final line is not checked twice because it is never checked in the loop because the code in the loop only checks a line when it precedes another and the final line does not precede another.