preserve empty next lines for line comments
when wrapping text to the next line, check whether the next line has content. if the next line is empty, do not merge text into it, because that destroys its emptiness  quality. instead, perform the wrap as if the next line does not exist, and insert a new line via replacement.

closes #46 at least for single line comments, will need to create another issue for block comments when i get to it