merge special case of last line of block into line logic in split
plumb all the handling of the special case for the last line of a block comment where the content is under the limit but the line is otherwise over the limit (e.g. because of the trailing whitespace and or the closing syntax) through the logic setup to split single line comments. now we have a more uniform approach and the logic is more organized. not all cases are handled yet so the code is still broken.