handle mismatched leading whitespace on split
if the next line has a different amount of leading whitespace, then the overflowing text from the current line should not be merged into the next line. a new line should be inserted instead. for javadoc comments, if the overflowing line is the first line of a block comment, then we tolerate 1 extra space on the next line when determining whether the text will be merged into the next line or just inserted.