trivial adjustment to merge
when checking whether the previous line overflows and therefore should not be considered for merging, prefer measuring the length of the line up to and including the close region of the comment, not just the suffix. even though the close region is pretty much always an empty string. this should result in no functional change