implement findIsolatedComments
break out the filtering logic from the main comment loop in index.ts' analyzeProgram function body. the loop is doing too much. it is incidental complexity. the helper function captures the preliminary step of excluding comments that we do not wrap at the moment. this might present difficulty in the future when adding support for such comments but I think it is better to worry about that later when the time comes, if ever.