impl isCandidateComment
deprecate findCandidateComments. figure out the awkward "this" syntax in typescript so that we can access extra content in isCandidateComment without the use of a closure.

using the isThing approach is more composable, the caller gets to use filter or whatever they want.

this approach reduces the sequential cohesion because we no longer care when this is called.